backend/requirements.txt
Django>=5.2
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.0
django-mongodb-backend
pymongo>=4.0
python-dotenv>=1.0
Pillow>=10.0
